refactor: split ContentView into Models, SidebarView, NavigationViews, DashboardViews
