+++ TimeScapeMac/TimeScape Planner Pro/TimeScape Planner Pro/SharedUIViews.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import AppKit
 import MapKit
 import Combine
 
@@ -436,3 +437,38 @@ import Combine
         }
     }
 
+
+    enum MicroFeedback {
+        static func tap() {
+            NSHapticFeedbackManager.defaultPerformer.perform(.alignment, performanceTime: .now)
+        }
+    }
+
+    extension Color {
+        init(hex: String) {
+            let raw = hex.trimmingCharacters(in: CharacterSet.alphanumerics.inverted)
+            var int: UInt64 = 0
+            Scanner(string: raw).scanHexInt64(&int)
+
+            let r: UInt64
+            let g: UInt64
+            let b: UInt64
+
+            switch raw.count {
+            case 3:
+                (r, g, b) = ((int >> 8) * 17, (int >> 4 & 0xF) * 17, (int & 0xF) * 17)
+            case 6:
+                (r, g, b) = (int >> 16, int >> 8 & 0xFF, int & 0xFF)
+            default:
+                (r, g, b) = (59, 130, 246)
+            }
+
+            self.init(
+                .sRGB,
+                red: Double(r) / 255,
+                green: Double(g) / 255,
+                blue: Double(b) / 255,
+                opacity: 1
+            )
+        }
+    }
